+++ packages/plugin-custom-enrichment-browser/rollup.config.js
@@ -1,0 +1,10 @@
+import { iife, umd } from '../../scripts/build/rollup.config';
+
+iife.input = umd.input;
+iife.output.name = 'amplitudeCustomEnrichmentPlugin';
+
+
+if (process.env.NODE_ENV === 'development') {
+  iife.output.sourcemap = 'inline';
+}
+export default [umd, iife];