+++ packages/plugin-page-url-enrichment-browser/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-url-enrichment-browser@0.6.12...@amplitude/plugin-page-url-enrichment-browser@0.7.0) (2026-03-20)
+
+
+### Features
+
+* **plugin-custom-enrichment:** add custom enrichment plugin ([#1339](https://github.com/amplitude/Amplitude-TypeScript/issues/1339)) ([b236190](https://github.com/amplitude/Amplitude-TypeScript/commit/b236190fa30d0e4325c37b8896824e11f980e20c))
+
+
+
+
+
+## [0.6.12](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-url-enrichment-browser@0.6.11...@amplitude/plugin-page-url-enrichment-browser@0.6.12) (2026-03-19)
+
+**Note:** Version bump only for package @amplitude/plugin-page-url-enrichment-browser
+
+
+
+
+
 ## [0.6.11](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-url-enrichment-browser@0.6.10...@amplitude/plugin-page-url-enrichment-browser@0.6.11) (2026-03-17)
 
 **Note:** Version bump only for package @amplitude/plugin-page-url-enrichment-browser